+++ RadixWallet/Core/DesignSystem/Colors.swift
@@ -272,3 +272,9 @@ extension Color {
 		)
 	}
 }
+
+extension Color {
+	var uiColor: UIColor {
+		.init(self)
+	}
+}
