+++ RadixWallet/Core/DesignSystem/Extensions/View+Extra.swift
@@ -19,26 +19,3 @@ extension View {
 		AnyView(self)
 	}
 }
-
-// MARK: - Apply
-extension View {
-	/// Applies `modifier` if given `condition` is met.
-	@ViewBuilder
-	func applyIf<Modifier>(_ condition: Bool, @ViewBuilder modifier: () -> Modifier) -> some View where Modifier: ViewModifier {
-		if condition {
-			self.modifier(modifier())
-		} else {
-			self
-		}
-	}
-
-	/// Applies `transform` if given `condition` is met.
-	@ViewBuilder
-	func applyIf<Content>(_ condition: Bool, @ViewBuilder transform: (Self) -> Content) -> some View where Content: View {
-		if condition {
-			transform(self)
-		} else {
-			self
-		}
-	}
-}
