+++ RadixWallet/Features/AppFeature/App+View.swift
@@ -1,30 +1,6 @@
 import ComposableArchitecture
 import SwiftUI
 
-extension View {
-	@MainActor
-	func configureNavigationBarAppearence() {
-		func setUp(appearence: UINavigationBarAppearance) {
-			appearence.titleTextAttributes = [.foregroundColor: UIColor(Color.app.gray1), .font: UIFont(font: FontFamily.IBMPlexSans.semiBold, size: 16)!]
-			appearence.largeTitleTextAttributes = [.foregroundColor: UIColor(Color.app.gray1)]
-			let image = UIImage(named: "arrow-back")
-			appearence.setBackIndicatorImage(image, transitionMaskImage: image)
-			appearence.backButtonAppearance.normal.titleTextAttributes = [.foregroundColor: UIColor.clear]
-		}
-
-		let scrollEdgeAppearence = UINavigationBarAppearance()
-		scrollEdgeAppearence.configureWithTransparentBackground()
-		setUp(appearence: scrollEdgeAppearence)
-		UINavigationBar.appearance().scrollEdgeAppearance = scrollEdgeAppearence
-
-		let standardAppearence = UINavigationBarAppearance()
-		standardAppearence.configureWithOpaqueBackground()
-		standardAppearence.backgroundColor = UIColor(Color.app.background)
-		setUp(appearence: standardAppearence)
-		UINavigationBar.appearance().standardAppearance = standardAppearence
-	}
-}
-
 // MARK: - App.View
 extension App {
 	@MainActor
