+++ RadixWallet/Features/AssetTransferFeature/Components/ChooseAccount/ChooseReceivingAccount+View.swift
@@ -158,7 +158,7 @@ private extension View {
 			action: ChooseReceivingAccount.Destination.Action.scanAccountAddress,
 			destination: {
 				ScanQRCoordinator.View(store: $0)
-					.setUpNavigationBar(title: L10n.AssetTransfer.ChooseReceivingAccount.scanQRNavigationTitle, includeBackground: false)
+					.setUpNavigationBar(title: L10n.AssetTransfer.ChooseReceivingAccount.scanQRNavigationTitle)
 			}
 		)
 	}
